# sauron-backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.CategoryWithRequiredStatusHierarchy:
    properties:
      child_categories:
        items:
          $ref: '#/definitions/handlers.CategoryWithRequiredStatusHierarchy'
        type: array
      created_at:
        description: Creation timestamp
        type: string
      description:
        description: Description of the category
        example: Core upper receiver components
        type: string
      id:
        description: Unique identifier for the category
        example: 1
        type: integer
      is_required:
        example: true
        type: boolean
      name:
        description: Name of the category
        example: Upper Assembly
        type: string
      parent_category:
        allOf:
        - $ref: '#/definitions/models.PartCategory'
        description: Parent category (self-referential relationship)
      parent_category_id:
        description: Parent category ID for hierarchical relationships (null for top-level
          categories)
        example: 0
        type: integer
      updated_at:
        description: Last update timestamp
        type: string
    type: object
...
  models.PartCategory:
    description: Hierarchical structure of part categories
    properties:
      child_categories:
        description: Child categories (reverse relationship)
        items:
          $ref: '#/definitions/models.PartCategory'
        type: array
      created_at:
        description: Creation timestamp
        type: string
      description:
        description: Description of the category
        example: Core upper receiver components
        type: string
      id:
        description: Unique identifier for the category
        example: 1
        type: integer
      name:
        description: Name of the category
        example: Upper Assembly
        type: string
      parent_category:
        allOf:
        - $ref: '#/definitions/models.PartCategory'
        description: Parent category (self-referential relationship)
      parent_category_id:
        description: Parent category ID for hierarchical relationships (null for top-level
          categories)
        example: 0
        type: integer
      updated_at:
        description: Last update timestamp
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@sauron.io
    name: Sauron API Support
    url: https://sauron.io/support
  description: A comprehensive API for firearms database management including parts,
    models, categories, product listings, and compatibility.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sauron.io/terms/
  title: Sauron Backend API
  version: "2.0"
paths:
...
  /firearm-models/{id}/categories:
    get:
      consumes:
      - application/json
      description: Retrieves all part categories associated with a specific firearm
        model
      parameters:
      - description: Firearm Model ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by required status (true=required, false=optional, omit=both)
        in: query
        name: required
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Array of part categories with required status
          schema:
            items:
              type: object
            type: array
        "400":
          description: Invalid firearm model ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Firearm model not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get part categories for a firearm model
      tags:
      - Firearm Models
      - Part Categories
  /firearm-models/{id}/categories-hierarchy:
    get:
      consumes:
      - application/json
      description: Retrieves all part categories associated with a specific firearm
        model in a hierarchical structure
      parameters:
      - description: Firearm Model ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Hierarchical array of part categories with required status
          schema:
            items:
              $ref: '#/definitions/handlers.CategoryWithRequiredStatusHierarchy'
            type: array
        "400":
          description: Invalid firearm model ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Firearm model not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get hierarchical categories for a firearm model
      tags:
      - Firearm Models
      - Part Categories
  /firearm-models/{id}/categories/{category_id}:
    delete:
      consumes:
      - application/json
      description: Removes the association between a part category and a firearm model
      parameters:
      - description: Firearm Model ID
        in: path
        name: id
        required: true
        type: integer
      - description: Part Category ID
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Relationship not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove a category from a firearm model
      tags:
      - Firearm Models
      - Part Categories
    post:
      consumes:
      - application/json
      description: Associates a part category with a firearm model
      parameters:
      - description: Firearm Model ID
        in: path
        name: id
        required: true
        type: integer
      - description: Part Category ID
        in: path
        name: category_id
        required: true
        type: integer
      - description: Relationship parameters
        in: body
        name: relationship
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Firearm model or part category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Relationship already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a category to a firearm model
      tags:
      - Firearm Models
      - Part Categories
    put:
      consumes:
      - application/json
      description: Updates the relationship between a firearm model and a part category
      parameters:
      - description: Firearm Model ID
        in: path
        name: id
        required: true
        type: integer
      - description: Part Category ID
        in: path
        name: category_id
        required: true
        type: integer
      - description: Updated relationship parameters
        in: body
        name: relationship
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Relationship not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a firearm model-category relationship
      tags:
      - Firearm Models
      - Part Categories
...
  /part-categories/firearm/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves all part categories associated with a specific firearm
        model
      parameters:
      - description: Firearm Model ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by required status (true=required, false=optional, omit=both)
        in: query
        name: required
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Array of part categories with required status
          schema:
            items:
              type: object
            type: array
        "400":
          description: Invalid firearm model ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Firearm model not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get part categories for a firearm model
      tags:
      - Firearm Models
      - Part Categories
...
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"